second commit fixed class data type int to long

+++ src/main/java/org/example/Pell.java
@@ -4,9 +4,9 @@ import java.util.Scanner;
 
 public class Pell {
     public static void main(String[] args) {
-        int result = 0;
-        int pn1 = 1;
-        int pn2;
+        long result = 0;
+        long pn1 = 1;
+        long pn2;
         try (Scanner scanner = new Scanner(System.in)) {
             int n = scanner.nextInt();
             for (int i = 1; i <= n; i++) {
